// src/commons/enums/http.enum.ts
'use strict';

export enum Message {
  INVALID_PARAMS = 'INVALID_PARAMS',
  INVALID_VALUE = 'INVALID_VALUE',
  INVALID_REQUEST = 'INVALID_REQUEST',
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  TOKEN_INVALID = 'TOKEN_INVALID',
  ROUTE_NOT_FOUND = 'ROUTE_NOT_FOUND',
  USER_LOGIN_ERROR = 'USER_LOGIN_ERROR',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
}

export enum StatusCode {
  OK = 200,
  BAD_REQUEST = 400,
  NOT_AUTHORIZED = 401,
  NOT_FOUND = 404,
  INTERNAL_ERROR = 500
}